requirements.txt
# MAMA Virtual Waiter - Dependencies

# Web Framework
streamlit>=1.28.0

# Environment
python-dotenv>=1.0.0

# HTTP Client
requests>=2.31.0

# Data Validation (optional)
pydantic>=2.0.0
